chore: bump pinned toolchain to 1.89.0 (#7919)
* chore: bump pinned toolchain to 1.89.0

Move the workspace off Rust 1.88.0 to 1.89.0.

- rust-toolchain.toml channel, workspace + fuzz + dhat-profile rust-version,
  the eight bench/compat workflows that hardcode the toolchain, the
  dependabot serial_test note, README, ONBOARDING, and the xtask preflight
  rust-version gate all move to 1.89.
- Satisfy 1.89 clippy: collapse nested ifs into let chains
  (clippy::collapsible_if now rewrites into stable let-chains under edition
  2024), replace &[x.clone()] with std::slice::from_ref, drop redundant
  same-type casts. Linux/Windows-only arms (io_uring, IOCP, statx, caps,
  SDDL, NetUserEnum) collapsed to match.
- 1.89 dead_code newly flags the validate `fuzzy` check, which was
  implemented in #6809 but never registered in checks::all(); wire it in
  alongside its siblings. Annotate the message-matcher `scenario` fields to
  match the existing #[allow(dead_code)] on their trait accessor.
- Apply rustfmt 1.89.

Historical/dated docs (changelog, dated audit and design snapshots) keep
their 1.88 references.

* chore: collapse nested ifs for the 1.89 clippy lint

* chore: reformat the post-rebase let-chain sites for 1.89 rustfmt

* chore: collapse the nested if in the wildmatch bytes property test

+++ crates/daemon/tests/connection_scaling_stress.rs
@@ -416,8 +416,8 @@ fn ensure_fd_capacity(required: u64) -> Result<u64, String> {
                 std::io::Error::last_os_error()
             ));
         }
-        let hard = rlim.rlim_max as u64;
-        let soft = rlim.rlim_cur as u64;
+        let hard = rlim.rlim_max;
+        let soft = rlim.rlim_cur;
         if required <= soft {
             return Ok(soft);
         }
